Set the ground sprite's entity.
Problem: The collision works only when the food is on the left side of the screen. I think the problem is caused by the sprite. I move the anchor point to fill the whole bottom of the screen instead of half. But that's not enough for the right side of the screen to handle the collision correctly.